refactor skeletons files
