+++ components/skeletons/player-card-skeleton.tsx
@@ -6,14 +6,14 @@ const PlayerCardSkeleton: NextPage = () => {
   return (
     <div className={styles.cardSkeleton}>
       <div className={styles.header}>
-        <div className={styles.left}></div>
-        <div className={styles.right}></div>
+        <div className={styles.left} />
+        <div className={styles.right} />
       </div>
       <div className={styles.body}>
-        <div className={styles.top}></div>
+        <div className={styles.top} />
         <div className={styles.bottom}>
-          <div className={styles.barr}></div>
-          <div className={styles.barr}></div>
+          <div className={styles.barr} />
+          <div className={styles.barr} />
         </div>
       </div>
     </div>
