+++ components/skeletons/player-stats-skeleton.module.scss
